refactor(contexts): reorganize filter context architecture with backward compatibility
Restructured filter context system by extracting implementations from
FilterActionsContext and FilterStateContext into separate modules while
maintaining backward compatibility through re-exports. Added PWA install
prompt component and updated service worker formatting.
