+++ dev-dist/registerSW.js
@@ -1,1 +1,2 @@
-if('serviceWorker' in navigator) navigator.serviceWorker.register('/dev-sw.js?dev-sw', { scope: '/', type: 'classic' })
+if ('serviceWorker' in navigator)
+  navigator.serviceWorker.register('/dev-sw.js?dev-sw', { scope: '/', type: 'classic' });
