+++ src/contexts/FilterProvider.tsx
@@ -5,8 +5,8 @@
  */
 
 import React, { ReactNode } from 'react';
-import { FilterStateProvider } from '@/contexts/FilterStateContext';
-import { FilterActionsProvider } from '@/contexts/FilterActionsContext';
+import { FilterStateProvider } from '@/contexts/FilterStateProvider';
+import { FilterActionsProvider } from '@/contexts/FilterActionsProvider';
 import { createLogger } from '@/utils/logger';
 
 const logger = createLogger('FilterProvider');
@@ -33,9 +33,20 @@ export const FilterProvider: React.FC<FilterProviderProps> = ({ children, user }
   );
 };
 
-/**
- * Export all consolidated hooks for easy importing
- */
-export * from '@/contexts/FilterHooks';
-export * from '@/contexts/FilterStateContext';
-export * from '@/contexts/FilterActionsContext';
+// Re-export essential hooks and types for backward compatibility
+export * from '@/contexts/filterState';
+export * from '@/contexts/filterActions';
+export { useFilterState } from '@/contexts/useFilterState';
+export { useFilterActions } from '@/contexts/useFilterActions';
+
+// Re-export FilterHooks for compatibility
+export {
+  useFilters,
+  useFilterStateOnly,
+  usePagination,
+  useSorting,
+  useFilterActionsOnly,
+  useStatusFilter,
+  useFiltersFull,
+  useFilterActionsAndMeta,
+} from '@/contexts/FilterHooks';
